+++ lib/src/classes/polygon.dart
@@ -64,6 +64,30 @@ class GeoJSONPolygon implements GeoJSONGeometry {
   @override
   double get distance => 0.0;
 
+  double get perimeter {
+    double calculateDistance(lat1, lon1, lat2, lon2) {
+      var p = 0.017453292519943295;
+      var c = cos;
+      var a = 0.5 -
+          c((lat2 - lat1) * p) / 2 +
+          c(lat1 * p) * c(lat2 * p) * (1 - c((lon2 - lon1) * p)) / 2;
+      return 12756.274 * asin(sqrt(a)) * 1000.0;
+    }
+
+    var perimeter = 0.0;
+    for (var i = 0; i < coordinates[0].length - 1; i++) {
+      var p1 = coordinates[0][i];
+      var p2 = coordinates[0][i + 1];
+      perimeter += calculateDistance(p1[1], p1[0], p2[1], p2[0]);
+    }
+    // Add distance between the last point and the first point to close the polygon
+    var p1 = coordinates[0][coordinates[0].length - 1];
+    var p2 = coordinates[0][0];
+    perimeter += calculateDistance(p1[1], p1[0], p2[1], p2[0]);
+
+    return perimeter;
+  }
+
   /// The constructor for the [coordinates] member
   GeoJSONPolygon(this.coordinates)
       : assert(coordinates.isNotEmpty,